fix circular dependency in BaseMessage class
