+++ src/lib/decorators/thread/ThreadDecorator.ts
@@ -1,6 +1,6 @@
 import { Expose } from 'class-transformer';
 import { Matches } from 'class-validator';
-import { MessageIdRegExp } from '../../agent/AgentMessage';
+import { MessageIdRegExp } from '../../agent/BaseMessage';
 
 /**
  * Represents `~thread` decorator
