+++ src/lib/decorators/thread/ThreadDecoratorExtension.ts
@@ -1,7 +1,7 @@
 import { Expose, Type } from 'class-transformer';
 import { ValidateNested } from 'class-validator';
 
-import { BaseMessageConstructor } from '../../agent/AgentMessage';
+import { BaseMessageConstructor } from '../../agent/BaseMessage';
 import { ThreadDecorator } from './ThreadDecorator';
 
 export function ThreadDecorated<T extends BaseMessageConstructor>(Base: T) {
