+++ src/lib/decorators/timing/TimingDecoratorExtension.ts
@@ -1,7 +1,7 @@
 import { Expose, Type } from 'class-transformer';
 import { ValidateNested } from 'class-validator';
 
-import { BaseMessageConstructor } from '../../agent/AgentMessage';
+import { BaseMessageConstructor } from '../../agent/BaseMessage';
 import { TimingDecorator } from './TimingDecorator';
 
 export function TimingDecorated<T extends BaseMessageConstructor>(Base: T) {
