+++ src/lib/decorators/transport/TransportDecoratorExtension.ts
@@ -1,7 +1,7 @@
 import { Expose, Type } from 'class-transformer';
 import { ValidateNested } from 'class-validator';
 import { TransportDecorator, ReturnRouteTypes } from './TransportDecorator';
-import { BaseMessageConstructor } from '../../agent/AgentMessage';
+import { BaseMessageConstructor } from '../../agent/BaseMessage';
 
 export function TransportDecorated<T extends BaseMessageConstructor>(Base: T) {
   class TransportDecoratorExtension extends Base {
