+++ src/lib/protocols/messagepickup/BatchMessage.ts
@@ -2,7 +2,8 @@ import { Equals, Matches, IsArray, ValidateNested } from 'class-validator';
 import { Type, Expose } from 'class-transformer';
 import uuid from 'uuid/v4';
 
-import { AgentMessage, MessageIdRegExp } from '../../agent/AgentMessage';
+import { MessageIdRegExp } from '../../agent/BaseMessage';
+import { AgentMessage } from '../../agent/AgentMessage';
 import { MessageType } from './messages';
 import { WireMessage } from '../../types';
